requirements.txt
# Python
python-dotenv

# Django
Django~=4.1
djangorestframework
django-filter
django-cors-headers
django-cleanup

# Miscroservices
dj-ms-auth-router==1.1.0

# Databases
psycopg2-binary
dj-database-url

# Deployment
gunicorn

# Cache
django-redis
redis

# Websocket support
channels[daphne]
channels-redis
uvicorn[standard]

# Async Celery tasks
celery
django-celery-beat

# Sentry
sentry-sdk

# App specific packages
Pillow